fix: Refactored Processor class into multiple smaller classes to improve maintenance
