+++ parsing/src/interfaces/Part.ts
@@ -1,0 +1,17 @@
+export interface Part {
+    name: string;
+    stackSize: number;
+    isFluid: boolean;
+    isFicsmas: boolean;
+}
+
+export interface PartDataInterface {
+    parts: { [key: string]: Part };
+    collectables: { [key: string]: string };
+    rawResources: { [key: string]: RawResource };
+}
+
+export interface RawResource {
+    name: string;
+    limit: number;
+}
